+++ square_constructors.cpp
@@ -1,3 +1,3 @@
 #include "square.hpp"
-Square::Square(point c_, float r_, Color color_) 
+Square::Square(point c_, int r_, Color color_) 
 	:Rectangle(c_, {r_,r_}, color_){};
